requirements.txt
streamlit
requests
azure-ai-inference